+++ packages/react-native-editor/CHANGELOG.md
@@ -10,6 +10,7 @@ For each user feature we should also add a importance categorization label  to i
 -->
 
 ## Unreleased
+-   [internal] Remove circular dependencies within the components package [#61102]
 
 ## 1.118.0
 -   [*] Fix a crash when pasting file images and special comment markup [#60476]